+++ test/TestLoanSourceRotation.t.sol
@@ -91,7 +91,7 @@ contract TestLoanSourceRotation is TestBaseWorkflow {
 
         FEE_PROJECT_ID = jbProjects().createFor(multisig());
 
-        SUCKER_REGISTRY = new JBSuckerRegistry(jbDirectory(), jbPermissions(), multisig(), address(0));
+        SUCKER_REGISTRY = new JBSuckerRegistry(jbDirectory(), jbPermissions(), jbPrices(), multisig(), address(0));
         HOOK_STORE = new JB721TiersHookStore();
         EXAMPLE_HOOK = new JB721TiersHook(
             jbDirectory(),