+++ test/TestTerminalEncodingInHash.t.sol
@@ -82,7 +82,7 @@ contract TestTerminalEncodingInHash is TestBaseWorkflow {
 
         FEE_PROJECT_ID = jbProjects().createFor(multisig());
 
-        SUCKER_REGISTRY = new JBSuckerRegistry(jbDirectory(), jbPermissions(), multisig(), address(0));
+        SUCKER_REGISTRY = new JBSuckerRegistry(jbDirectory(), jbPermissions(), jbPrices(), multisig(), address(0));
         HOOK_STORE = new JB721TiersHookStore();
         JB721TiersHook exampleHook = new JB721TiersHook(
             jbDirectory(),